# [Perf] LiteLLM Proxy:  +400 RPS when using correct amount of CPU cores 