## [Perf] +400 RPS when using correct amount of CPU cores 

## Problem:
On deployed machines LiteLLM was not using 100% of the CPU, infact LiteLLM was using 0.1% of the CPU - as a result we were not maximizing the throughput of LiteLLM Proxy 

## Solution 
Update `DEFAULT_NUM_WORKERS_LITELLM_PROXY=4` - this ensures we use 1 worker per CPU core. (LiteLLM instructs teams to deploy with 4 vCPUs, so 4 vCPUs with 1 worker) 

## Performance Test Results
Roughly a 166% RPs increase 

### Baseline (384 RPS) - Before this PR

| Type | Name | Median (ms) | 95%ile (ms) | 99%ile (ms) | Average (ms) | Current RPS |
|------|------|-------------|-------------|-------------|--------------|-------------|
| POST | /chat/completions | 1300 | 2600 | 15000 | 1634.55 | 384.1 |
| **Aggregated** | | **1300** | **2600** | **15000** | **1634.55** | **384.1** |

### Using 100% CPU - with 4 workers. Workers == num CPU cores 

| Type | Name | Median (ms) | 95%ile (ms) | 99%ile (ms) | Average (ms) | Current RPS |
|------|------|-------------|-------------|-------------|--------------|-------------|
| POST | /chat/completions | 160 | 610 | 750 | 229.64 | 1022.3 |
| **Aggregated** | | **160** | **610** | **750** | **229.64** | **1022.3** |



## Some more technical details on why this change 

I was able to confirm we had good single-thread latency, but under high traffic cpu usage was not even at 1% - this indicated underutilization of cores on machines. The main bottleneck was in how we were using the available hardware. 


<!-- e.g. "Implement user authentication feature" -->

## Relevant issues

<!-- e.g. "Fixes #000" -->

## Pre-Submission checklist

**Please complete all items before asking a LiteLLM maintainer to review your PR**

- [x] I have Added testing in the [`tests/litellm/`](https://github.com/BerriAI/litellm/tree/main/tests/litellm) directory, **Adding at least 1 test is a hard requirement** - [see details](https://docs.litellm.ai/docs/extras/contributing_code)
- [x] I have added a screenshot of my new test passing locally 
- [x] My PR passes all unit tests on [`make test-unit`](https://docs.litellm.ai/docs/extras/contributing_code)
- [x] My PR's scope is as isolated as possible, it only solves 1 specific problem


## Type

<!-- Select the type of Pull Request -->
<!-- Keep only the necessary ones -->

🆕 New Feature
🚄 Infrastructure
✅ Test

## Changes